refactor: decompose ProductHub.tsx into sub-components
Extract 6 presentation components from ProductHub (2,716 -> 2,081 lines):
- SortableProductCard: drag-and-drop product card with Pick2 fields
- DragOverlayItem: drag overlay ghost element
- DroppableColumn: droppable lane wrapper
- DuplicatesPanel: duplicate detection display
- RecommendedPackagePanel: recommended package radio group
- FilterBar: package lane, publish, featured, category filters
- BulkActionsBar: bulk publish/unpublish/featured/category controls

State management stays in ProductHub as the orchestrator.
Pick2 config panel left in ProductHub due to deep state entanglement.
All 297 tests passing, build clean.
