+++ src/components/product-hub/index.ts
@@ -1,0 +1,9 @@
+export { SortableProductCard } from "./SortableProductCard";
+export type { SortableProductCardProps } from "./SortableProductCard";
+export { DragOverlayItem } from "./DragOverlayItem";
+export { DroppableColumn } from "./DroppableColumn";
+export { DuplicatesPanel } from "./DuplicatesPanel";
+export type { DuplicateGroup } from "./DuplicatesPanel";
+export { RecommendedPackagePanel } from "./RecommendedPackagePanel";
+export { FilterBar } from "./FilterBar";
+export { BulkActionsBar } from "./BulkActionsBar";
